// frontend/src/api/generated/models/PortfolioPositionMutationData.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { PortfolioPositionDTO } from './PortfolioPositionDTO';
export type PortfolioPositionMutationData = {
    position: PortfolioPositionDTO;
    portfolio_revision: number;
};