// frontend/src/api/generated/models/PortfolioPositionsData.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { PortfolioPositionDTO } from './PortfolioPositionDTO';
export type PortfolioPositionsData = {
    portfolio_id: string;
    portfolio_revision: number;
    items: Array<PortfolioPositionDTO>;
};