// frontend/src/api/generated/models/WatchlistItemMutationData.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { WatchlistItemDTO } from './WatchlistItemDTO';
export type WatchlistItemMutationData = {
    item: (WatchlistItemDTO | null);
    watchlist_revision: number;
};