// frontend/src/api/generated/models/WatchlistItemOrderUpdateRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { WatchlistItemOrderEntry } from './WatchlistItemOrderEntry';
export type WatchlistItemOrderUpdateRequest = {
    expected_watchlist_revision: number;
    items: Array<WatchlistItemOrderEntry>;
};